+++ src/main/java/org/backend/controllers/PostController.java
@@ -3,7 +3,7 @@ package org.backend.controllers;
 import org.backend.Main;
 import org.backend.config.Config;
 import org.backend.data.DataSaverAndLoader;
-import org.backend.employee.PostEmployee;
+import org.backend.models.PostEmployee;
 import lombok.NonNull;
 
 import java.nio.file.Path;
