+++ src/main/java/org/backend/files/FileDataBuilder.java
@@ -3,9 +3,9 @@ package org.backend.files;
 import org.backend.Main;
 import org.backend.config.Config;
 import org.backend.data.DataBuilder;
-import org.backend.employee.BaseMethodsForEmployeeAndPost;
-import org.backend.employee.Employee;
-import org.backend.employee.PostEmployee;
+import org.backend.models.BaseMethodsForEmployeeAndPost;
+import org.backend.models.Employee;
+import org.backend.models.PostEmployee;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NonNull;
