+++ src/test/java/TestFileBuilder.java
@@ -2,8 +2,8 @@ import org.backend.Main;
 import org.backend.controllers.PostController;
 import static org.mockito.Mockito.*;
 import org.backend.data.DataBuilder;
-import org.backend.employee.Employee;
-import org.backend.employee.PostEmployee;
+import org.backend.models.Employee;
+import org.backend.models.PostEmployee;
 import org.backend.files.FileDataBuilder;
 import org.assertj.core.internal.bytebuddy.utility.RandomString;
 import static org.assertj.core.api.Assertions.*;
