+++ src/test/java/TestFileSaverAndLoader.java
@@ -4,8 +4,8 @@ import org.backend.controllers.EmployeeController;
 import org.backend.controllers.PostController;
 import org.backend.data.DataBuilder;
 import org.backend.data.DataSaverAndLoader;
-import org.backend.employee.Employee;
-import org.backend.employee.PostEmployee;
+import org.backend.models.Employee;
+import org.backend.models.PostEmployee;
 import org.backend.files.FileDataBuilder;
 import org.backend.files.FileSaverAndLoader;
 import org.assertj.core.internal.bytebuddy.utility.RandomString;
