+++ src/test/java/TestFindObjects.java
@@ -2,8 +2,8 @@ import org.assertj.core.internal.bytebuddy.utility.RandomString;
 import org.backend.Main;
 import org.backend.controllers.EmployeeController;
 import org.backend.controllers.PostController;
-import org.backend.employee.Employee;
-import org.backend.employee.PostEmployee;
+import org.backend.models.Employee;
+import org.backend.models.PostEmployee;
 import org.backend.files.FileDataBuilder;
 import org.backend.files.FileSaverAndLoader;
 import org.backend.observer.Observer;
