+++ src/test/java/TestJsonBuilder.java
@@ -1,12 +1,11 @@
 import com.google.gson.Gson;
 import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
 import org.assertj.core.internal.bytebuddy.utility.RandomString;
 import org.backend.Main;
 import org.backend.controllers.PostController;
 import org.backend.data.DataBuilder;
-import org.backend.employee.Employee;
-import org.backend.employee.PostEmployee;
+import org.backend.models.Employee;
+import org.backend.models.PostEmployee;
 import org.backend.json.JsonBuilder;
 import org.backend.utils.CheckData;
 import org.junit.jupiter.api.BeforeAll;
